Update config to 1.16

+++ src/main/java/team/cqr/cqrepoured/item/IExtendedItemTier.java
@@ -4,8 +4,10 @@ import net.minecraft.item.IItemTier;
 
 public interface IExtendedItemTier extends IItemTier {
 
-	public int getFixedAttackDamageBonus();
+	public float getFixedAttackDamageBonus();
+
 	public float getAttackSpeedBonus();
+
 	public double getMovementSpeedBonus();
-	
+
 }
